Turn on linux tests for darwin, too
