+++ chown_stub.go
@@ -1,0 +1,11 @@
+// +build !linux,!darwin
+
+package lumberjack
+
+import (
+	"os"
+)
+
+func chown(_ string, _ os.FileInfo) error {
+	return nil
+}
